REFACTOR: Moved Hibernate Search specific files into own package
