+++ src/main/java/mops/module/database/Modul.java
@@ -13,7 +13,7 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import lombok.Getter;
 import lombok.Setter;
-import mops.module.services.IndexWhenVisibleInterceptor;
+import mops.module.searchconfig.IndexWhenVisibleInterceptor;
 import org.hibernate.search.annotations.Field;
 import org.hibernate.search.annotations.Indexed;
 import org.hibernate.search.annotations.IndexedEmbedded;
